POJO-Support: Compel knockout to generate _ko_property_writers for a few more of its bindings, by marking them as two-way.
Some of knockout's built-in bindings (checked, hasFocus, selectedOptions, textInput, and value) are already listed as "two-way bindings" by default. For these, knockout generates an additional binding named "_ko_property_writers" which exposes functions for writing these values to non-observable model properties.

We need to tell knockout to generate property writers for more of its built-in bindings, so that `init` can use these to write to the model.